+++ src/Package/C/Type.hs
@@ -39,6 +39,7 @@ data Command = CreateDirectory { dir :: String }
              | Symlink { tgt :: String, linkName :: String }
              | Write { contents :: T.Text, file :: FilePath }
              | CopyFile { src :: FilePath, dest :: FilePath }
+             | Patch { patchContents :: T.Text }
              deriving (Eq, Ord, Generic, Binary, Hashable)
 
 -- TODO: build script should take OS as an argument?
@@ -66,6 +67,7 @@ commandDhallToCommand (Dhall.SymlinkBinary b)     = SymlinkBinary (T.unpack b)
 commandDhallToCommand (Dhall.Write out fp)        = Write out (T.unpack fp)
 commandDhallToCommand (Dhall.CopyFile src' dest') = CopyFile (T.unpack src') (T.unpack dest')
 commandDhallToCommand (Dhall.Symlink t l)         = Symlink (T.unpack t) (T.unpack l)
+commandDhallToCommand (Dhall.Patch c)             = Patch c
 
 buildVarsToDhallBuildVars :: BuildVars -> Dhall.BuildVars
 buildVarsToDhallBuildVars (BuildVars dir' cd tgt' cross incls prelds shr lds bins os' arch' sta nproc) = Dhall.BuildVars (T.pack dir') (T.pack cd) tgt' cross (T.pack <$> incls) (T.pack <$> prelds) (T.pack <$> shr) (T.pack <$> lds) (T.pack <$> bins) os' arch' sta (fromIntegral nproc)