+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # cpkg
 
+## 0.1.3.1
+
+  * Don't install build-tool dependencies globally
+
+## 0.1.3.0
+
+  * Add ability to patch libraries
+
+## 0.1.2.1
+
+  * Add `--global` flag
+
 ## 0.1.2.0
 
   * Export `Dep`